refactor: restructured the folder directory
