+++ src/Utils/Objects/StaticImage.ts
