+++ src/Utils/Objects/StaticText.ts
