+++ src/Utils/Objects/StaticVector.ts
